# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: PredictIQ API
  version: 1.0.0
  description: |
    REST API for the PredictIQ prediction markets platform.

    ## API Versioning

    The API uses URL path versioning (`/api/v1/`). The current stable version is **v1**.

    Clients may also send an `API-Version` header (e.g. `API-Version: v1`) to explicitly
    declare the version they target. If omitted, the server defaults to the current version.

    ## Deprecation Policy

    When a version is deprecated:
    - Responses will include a `Deprecation` header set to `true`.
    - A `Sunset` header will indicate the date after which the version will be removed.
    - A `Link` header will point to migration documentation.

    Clients should monitor these headers and migrate before the sunset date.

    Deprecated versions are supported for a minimum of **12 months** after the deprecation
    announcement before being removed.
servers:
  - url: http://0.0.0.0:8080
    description: Local development (default bind 0.0.0.0:8080)
...
paths:
...
  /api/v1/statistics:
    get:
      tags: [markets]
      operationId: getStatistics
      summary: Platform statistics
      parameters:
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Statistics payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/markets/featured:
    get:
      tags: [markets]
      operationId: getFeaturedMarkets
      summary: List featured markets
      parameters:
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Array of featured markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeaturedMarketView"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/content:
    get:
      tags: [markets]
      operationId: getContent
      summary: Paginated content feed
      parameters:
        - $ref: "#/components/parameters/apiVersion"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Content payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/markets/{market_id}/resolve:
    post:
      tags: [markets]
      operationId: resolveMarket
      summary: Resolve a market and invalidate caches (admin)
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/marketId"
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Cache invalidation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidationResult"
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "500":
          $ref: "#/components/responses/ApiError"
...
  /api/v1/blockchain/markets/{market_id}:
    get:
      tags: [blockchain]
      operationId: getBlockchainMarket
      summary: On-chain market data
      parameters:
        - $ref: "#/components/parameters/marketId"
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Market chain data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/blockchain/stats:
    get:
      tags: [blockchain]
      operationId: getBlockchainStats
      summary: Platform on-chain statistics
      parameters:
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/blockchain/users/{user}/bets:
    get:
      tags: [blockchain]
      operationId: getUserBets
      summary: Bets placed by a user
      parameters:
        - name: user
          in: path
          required: true
          schema:
            type: string
          description: Stellar address of the user
        - $ref: "#/components/parameters/apiVersion"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: User bets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/blockchain/oracle/{market_id}:
    get:
      tags: [blockchain]
      operationId: getOracleResult
      summary: Oracle result for a market
      parameters:
        - $ref: "#/components/parameters/marketId"
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Oracle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/blockchain/tx/{tx_hash}:
    get:
      tags: [blockchain]
      operationId: getTransactionStatus
      summary: Transaction status by hash
      parameters:
        - name: tx_hash
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Transaction status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/newsletter/subscribe:
    post:
      tags: [newsletter]
      operationId: newsletterSubscribe
      summary: Subscribe to newsletter
      parameters:
        - $ref: "#/components/parameters/apiVersion"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeRequest"
      responses:
        "200":
          $ref: "#/components/responses/NewsletterResponse"
        "400":
          $ref: "#/components/responses/NewsletterResponse"
        "409":
          $ref: "#/components/responses/NewsletterResponse"
        "429":
          $ref: "#/components/responses/NewsletterResponse"
  /api/v1/newsletter/confirm:
    get:
      tags: [newsletter]
      operationId: newsletterConfirm
      summary: Confirm newsletter subscription via token
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/NewsletterResponse"
        "400":
          $ref: "#/components/responses/NewsletterResponse"
        "404":
          $ref: "#/components/responses/NewsletterResponse"
  /api/v1/newsletter/unsubscribe:
    delete:
      tags: [newsletter]
      operationId: newsletterUnsubscribe
      summary: Unsubscribe from newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          $ref: "#/components/responses/NewsletterResponse"
        "400":
          $ref: "#/components/responses/NewsletterResponse"
  /api/v1/newsletter/gdpr/export:
    get:
      tags: [newsletter]
      operationId: newsletterGdprExport
      summary: GDPR data export for a subscriber
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Subscriber data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterExportResponse"
        "400":
          $ref: "#/components/responses/NewsletterResponse"
        "404":
          $ref: "#/components/responses/NewsletterResponse"
  /api/v1/newsletter/gdpr/delete:
    delete:
      tags: [newsletter]
      operationId: newsletterGdprDelete
      summary: GDPR data deletion for a subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          $ref: "#/components/responses/NewsletterResponse"
        "400":
          $ref: "#/components/responses/NewsletterResponse"
  /api/v1/email/preview/{template_name}:
    get:
      tags: [email]
      operationId: emailPreview
      summary: Preview a rendered email template (admin)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: template_name
          in: path
          required: true
          schema:
            type: string
            enum:
              - newsletter_confirmation
              - waitlist_confirmation
              - contact_form_auto_response
              - welcome_email
      responses:
        "200":
          description: Rendered email preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/email/test:
    post:
      tags: [email]
      operationId: emailSendTest
      summary: Send a test email (admin)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTestRequest"
      responses:
        "200":
          description: Test email result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTestResponse"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/email/analytics:
    get:
      tags: [email]
      operationId: getEmailAnalytics
      summary: Email analytics (admin)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: template_name
          in: query
          required: false
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /api/v1/email/queue/stats:
    get:
      tags: [email]
      operationId: getEmailQueueStats
      summary: Email queue statistics (admin)
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Queue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyObject"
        "500":
          $ref: "#/components/responses/ApiError"
  /webhooks/sendgrid:
    post:
      tags: [webhooks]
      operationId: sendgridWebhook
      summary: SendGrid event webhook receiver
      security:
        - SendGridWebhook: []
      parameters:
        - name: X-Twilio-Email-Event-Webhook-Signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC-SHA256 signature of timestamp+body, base64-encoded
        - name: X-Twilio-Email-Event-Webhook-Timestamp
          in: header
          required: true
          schema:
            type: string
          description: Unix timestamp of the request (replay protection window is 300 s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AnyObject"
      responses:
        "200":
          description: Webhook processed
        "401":
          $ref: "#/components/responses/ApiError"
components:
  parameters:
    apiVersion:
      name: API-Version
      in: header
      required: false
      schema:
        type: string
        enum: [v1]
        default: v1
      description: |
        Target API version. Defaults to the current stable version (`v1`).
        Responses for deprecated versions include `Deprecation`, `Sunset`, and `Link` headers.
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        Client-generated unique key (e.g. UUID v4) to deduplicate mutating requests.
        If a request with the same key was processed within the idempotency window,
        the cached response is returned immediately without re-executing the operation.
    marketId:
      name: market_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
        default: 1
    pageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
    FeaturedMarketView:
      type: object
      required: [id, title, volume, ends_at, onchain_volume]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        volume:
          type: number
          format: double
        ends_at:
          type: string
          format: date-time
        onchain_volume:
          type: string
        resolved_outcome:
          type: integer
          format: int32
          nullable: true
    InvalidationResult:
      type: object
      required: [invalidated_keys]
      properties:
        invalidated_keys:
          type: integer
    NewsletterSubscribeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        source:
          type: string
          maxLength: 64
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    NewsletterResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    NewsletterExportResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AnyObject"
    EmailTestRequest:
      type: object
      required: [recipient, template_name]
      properties:
        recipient:
          type: string
          format: email
        template_name:
          type: string
    EmailTestResponse:
      type: object
      required: [success, message, message_id]
      properties:
        success:
          type: boolean
        message:
          type: string
        message_id:
          type: string
  responses:
    ApiError:
      description: Error response with machine-readable code
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NewsletterResponse:
      description: Newsletter operation result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NewsletterResponse"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key required for protected endpoints
    SendGridWebhook:
      type: apiKey
      in: header
      name: X-Twilio-Email-Event-Webhook-Signature
      description: HMAC-SHA256 signature provided by SendGrid on webhook delivery